Additional warning fixes; run brew update before brew install for hamlib

+++ src/process/SpectrumVisualProcessor.cpp
@@ -286,7 +286,7 @@ void SpectrumVisualProcessor::process() {
                 return;
             }
             
-            while (resampleBw / SPECTRUM_VZM >= bandwidth) {
+            while (resampleBw / SPECTRUM_VZM >= (long) bandwidth) {
                 resampleBw /= SPECTRUM_VZM;
             }
